Split xds authentication
